# docs/swagger.yaml
openapi: 3.1.0
info:
  title: Api-Symfony-Shell
  version: 1.1.0
servers:
  - url: http://localhost:8000
components:
  securitySchemes:
    bearerCookie:
      type: apiKey
      in: cookie
      name: BEARER
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, details]
      properties:
        code:
          type: string
          examples: [bad_request, unauthorized, access_denied, not_found, conflict, rate_limited, internal_error]
        message:
          type: string
        details:
          nullable: true
          oneOf:
            - type: string
            - type: object
            - type: array
...
paths:
...
  /api/public/users/me:
    get:
      security:
        - bearerCookie: []
      responses:
        '200':
          description: Current public user
    patch:
      security:
        - bearerCookie: []
      responses:
        '200':
          description: Current public user updated
        '403':
          description: CSRF mismatch
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'